requirements.txt
aiogram
python-dotenv
bs4
sqlalchemy
asyncpg
